Move `RoundingIncrement` to its own file
